// src/app/api/fq/fq000101/fq000101-res.ts
import { ResBody } from '@base/api/model/res-body';

export class FQ000101ResBody extends ResBody {
    trnsRsltCode: string;
    hostCode: string;
    hostCodeMsg: string;
    isAgreeQRCode: string;
    defaultTrnsOutAcct: string;
    trnsLimitAmt: string;
    mobileBarcode: string;
    loveCode: string;
    socialWelfareName: string;
    defaultBarcode: string;
    defaultTrnsCard: string;
    trnsOutAccts: any;
}